api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "personal-dashboard-api"
version = "0.1.0"
description = "Personal Dashboard API using FastAPI"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlmodel>=0.0.16",
    "python-dotenv>=1.0.1",
    "pydantic[email]>=2.6.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    # PostgreSQL支持（生产环境）
    "psycopg2-binary>=2.9.9; python_version>='3.8' and platform_system!='Windows'",
    "psycopg2>=2.9.9; platform_system=='Windows'",
]
requires-python = ">=3.11"

[project.urls]
Homepage = "https://github.com/yourusername/personal-dashboard"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]